Clarify CHK raft safety layout and soft-requeue deferred quorum waits.
Move ensemble policy into worker-raft-safety, collapse the disrupt gate behind
ensureQuorumSafeToDisruptHost, and requeue ErrCRUDDeferred after 5s instead of
error backoff so Raft headroom waits stay intentional.
